- Changed all doubles to a new decmial class (for speed). Be aware   that the class can only handle numbers in the low trillions; it's   technically possible to overflow for a large dataset, but we've   never yet seen numbers big enough in practice.

+++ include/workload/common/q6_aggregate.h
@@ -15,7 +15,7 @@ using namespace qpipe;
 class q6_count_aggregate_t : public tuple_aggregate_t {
 
     struct agg_t {
-        double sum;
+        decimal sum;
         int count;
     };
 
@@ -32,7 +32,7 @@ public:
     
     virtual void aggregate(char* agg_data, const tuple_t & src) {
         agg_t* agg = aligned_cast<agg_t>(agg_data);
-	double * d = aligned_cast<double>(src.data);
+	decimal * d = aligned_cast<decimal>(src.data);
         
 	// update COUNT and SUM
 	agg->count++;
@@ -46,7 +46,7 @@ public:
 
     virtual void finish(tuple_t &dest, const char* agg_data) {
         agg_t* agg = aligned_cast<agg_t>(agg_data);
-        double *output = aligned_cast<double>(dest.data);
+        decimal *output = aligned_cast<decimal>(dest.data);
         output[0] = agg->sum;
         output[1] = agg->count;
     }
